+++ test/cases/require-fonts/webpack.config.js
@@ -1,0 +1,49 @@
+const path = require('path');
+const PugPlugin = require('pug-plugin');
+
+module.exports = {
+  stats: {
+    children: true,
+  },
+
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+  },
+
+  entry: {
+    index: 'src/index.pug',
+  },
+
+  plugins: [new PugPlugin()],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: 'pug-loader',
+        options: {
+          method: 'render',
+        },
+      },
+
+      {
+        test: /\.(png|jpg|jpeg)/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/images/[name].[hash:8][ext]',
+        },
+      },
+
+      {
+        test: /\.(eot|ttf|woff|woff2)/,
+        type: 'asset/resource',
+        generator: {
+          filename: 'assets/fonts/[name][ext]',
+        },
+      },
+    ],
+  },
+};