+++ test/utils/file.js
@@ -5,24 +5,24 @@ const { execSync } = require('child_process');
 /**
  * Get files with relative paths.
  * @param {string} dir
+ * @param {boolean} returnAbsolutePath If is false then return relative paths by dir.
  * @return {[]}
  */
-export const readDirRecursiveSync = function (dir = './') {
+export const readDirRecursiveSync = function (dir = './', returnAbsolutePath = true) {
   const entries = fs.readdirSync(dir, { withFileTypes: true });
 
   dir = path.resolve(dir);
 
-  // Get files within the current directory and add a path key to the file objects
+  // get files within the current directory and add a path key to the file objects
   const files = entries.filter((file) => !file.isDirectory()).map((file) => path.join(dir, file.name));
-
-  // Get folders within the current directory
+  // get folders within the current directory
   const folders = entries.filter((folder) => folder.isDirectory());
 
   for (const folder of folders) {
     files.push(...readDirRecursiveSync(path.join(dir, folder.name)));
   }
 
-  return files;
+  return returnAbsolutePath ? files : files.map((file) => file.replace(path.join(dir, '/'), ''));
 };
 
 /**
@@ -47,17 +47,6 @@ export const copyRecursiveSync = function (src, dest) {
 };
 
 /**
- * Return object of JSON file.
- *
- * @param {string} file
- * @return {any}
- */
-export const readJsonSync = (file) => {
-  const content = fs.readFileSync(file, 'utf-8');
-  return JSON.parse(content);
-};
-
-/**
  * Return content of file as string.
  *
  * @param {string} file
@@ -65,7 +54,7 @@ export const readJsonSync = (file) => {
  */
 export const readTextFileSync = (file) => {
   if (!fs.existsSync(file)) {
-    console.log(`WARN: the file "${file}" not found.`);
+    console.log(`\nWARN: the file "${file}" not found.`);
     return '';
   }
   return fs.readFileSync(file, 'utf-8');
@@ -81,17 +70,4 @@ export const execScriptSync = (file) => {
   const result = execSync('node ' + file);
   // replace last newline in result
   return result.toString().replace(/\n$/, '');
-};
-
-export const getCompareFiles = function (PATHS, absTestPath, file, manifest) {
-  let sourceFile = path.join(PATHS.output, file);
-  let expectedFile = path.join(PATHS.expected, file);
-  let targetFile = manifest[sourceFile] || sourceFile;
-  let received = readTextFileSync(path.join(absTestPath, PATHS.webRoot, targetFile));
-  let expected = readTextFileSync(path.join(absTestPath, expectedFile));
-
-  return {
-    received: received,
-    expected: expected,
-  };
 };