+++ test/utils/helpers.js
@@ -1,11 +1,46 @@
 import path from 'path';
-import { execScriptSync, readTextFileSync } from './file';
+import { readDirRecursiveSync, readTextFileSync, execScriptSync } from './file';
 import { compile } from './webpack';
+
+/**
+ * This is the patch for some environments, like `jest`.
+ * The `jest` hasn't in global scope the `btoa` function which used in `css-loader`.
+ */
+if (typeof global.btoa === 'undefined') {
+  global.btoa = (input) => Buffer.from(input, 'latin1').toString('base64');
+}
+
+export const getCompareFileList = function (receivedPath, expectedPath) {
+  return {
+    received: readDirRecursiveSync(receivedPath, false).sort(),
+    expected: readDirRecursiveSync(expectedPath, false).sort(),
+  };
+};
 
 export const getCompareFileContents = function (receivedFile, expectedFile, filter = /.(html|css|css.map)$/) {
   return filter.test(receivedFile) && filter.test(expectedFile)
     ? { received: readTextFileSync(receivedFile), expected: readTextFileSync(expectedFile) }
     : { received: '', expected: '' };
+};
+
+export const compareFileListAndContent = (PATHS, relTestCasePath, done) => {
+  const absTestPath = path.join(PATHS.testOutput, relTestCasePath),
+    webRootPath = path.join(absTestPath, PATHS.webRoot),
+    expectedPath = path.join(absTestPath, PATHS.expected);
+
+  compile(PATHS, relTestCasePath, {}).then(() => {
+    const { received: receivedFiles, expected: expectedFiles } = getCompareFileList(webRootPath, expectedPath);
+    expect(receivedFiles).toEqual(expectedFiles);
+
+    expectedFiles.forEach((file) => {
+      const { received, expected } = getCompareFileContents(
+        path.join(webRootPath, file),
+        path.join(expectedPath, file)
+      );
+      expect(received).toEqual(expected);
+    });
+    done();
+  });
 };
 
 export const compareContent = (